✨ Tested 1 SP on short path, 2 SPs on long path
According to profiling results, even the short path needs a scheduling point(#1) halfway. #1 is at the end of TCP. Long path needs another one in remaining processing(#2). #2 is after the first rebuilt is done. The long path needs to stall during the short one, so I add an interface to apply per-stack lock, such that it will not be executed until the completion of short path unlock it. This needs another SP(#3). #3 is inside stream module, which is very early.\n Implementation Notes: 1. Context complete follows a strict FIFO model, so it might have problems if the rebuilt path is indeed too long.\n 2. Currently I assume a path can be identified before actual branch, which might not hold in really corner cases.